+++ alicloud/data_source_alicloud_vswitches.go
@@ -101,13 +101,16 @@ func dataSourceAlicloudVSwitchesRead(d *schema.ResourceData, meta interface{}) e
 
 	args := vpc.CreateDescribeVSwitchesRequest()
 	args.RegionId = string(client.Region)
-	args.PageSize = requests.NewInteger(PageSizeLarge)
+	// API DescribeVSwitches has some limitations
+	// If there is no vpc_id, setting PageSizeSmall can avoid ServiceUnavailable Error
+	args.PageSize = requests.NewInteger(PageSizeSmall)
 	args.PageNumber = requests.NewInteger(1)
 	if v, ok := d.GetOk("zone_id"); ok {
 		args.ZoneId = Trim(v.(string))
 	}
 	if v, ok := d.GetOk("vpc_id"); ok {
 		args.VpcId = Trim(v.(string))
+		args.PageSize = requests.NewInteger(PageSizeLarge)
 	}
 
 	var allVSwitches []vpc.VSwitch